+++ CHANGELOG.md
@@ -19,6 +19,7 @@
    * CHANGED: Optimize `get_service_days` in servicedays.cc [#5952](https://github.com/valhalla/valhalla/pull/5952)
    * ADDED: OSM XML format support for tile building [#5934](https://github.com/valhalla/valhalla/pull/5934)
    * ADDED: access restriction layer in MVT [#5912](https://github.com/valhalla/valhalla/pull/5912)
+   * ADDED: incidents to locate JSON response [#5968](https://github.com/valhalla/valhalla/pull/5968)
 
 ## Release Date: 2026-02-19 Valhalla 3.6.3
 * **Removed**